refactor NameSpacing utilities to own file...
breaking compile link to `.vue` files
