+++ packages/vue/src/components/Edit/ViewableDataInputForm/DataInputForm.vue
@@ -95,7 +95,8 @@ import DataShapeTable from '@/components/Edit/DataTable/DataShapeTable.vue';
 import TagsInput from '@/components/Edit/TagsInput.vue';
 import { FieldInput } from '@/components/Edit/ViewableDataInputForm/FieldInput';
 import { alertUser } from '@/components/SnackbarService.vue';
-import Courses, { NameSpacer, ShapeDescriptor } from '@/courses';
+import Courses from '@/courses';
+import { NameSpacer, ShapeDescriptor } from '@/courses/NameSpacer';
 import { fieldConverters, FieldType } from '@/enums/FieldType';
 import { Status } from '@/enums/Status';
 import { CourseConfig } from '@/server/types';
